# Conflicts:
#	src/main/java/com/cgv/s1/oproduct/OproductDTO.java
#	src/main/java/com/cgv/s1/oproduct/review/OproductReviewService.java
#	src/main/java/com/cgv/s1/util/Pager.java
#	src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
#	src/main/webapp/WEB-INF/views/common/productReview.jsp
#	src/test/java/com/cgv/s1/oproduct/OproductDAOTest.java

+++ src/main/java/com/cgv/s1/oproduct/OproductDTO.java
@@ -15,12 +15,15 @@ public class OproductDTO {
 	private String productName;
 	private String productDetail;
 	private Integer productPrice;
-	//퍼센트 double에서 integer로 바꿈
-	private Double productDC;
+
+	//퍼센트 double에서 integer로 바꿈 -> 다시 double로 바꿈(04.08) -> 다시 integer
+	private Integer productDC;
+
 	private Integer productStock;
 	private Integer productSale;
 	private Date regDate;
 	
+
 	//디테일 파일 담기 
 	private List<OproductDTO> fileDTOs;
 	
@@ -98,10 +101,10 @@ public class OproductDTO {
 	public void setProductPrice(Integer productPrice) {
 		this.productPrice = productPrice;
 	}
-	public Double getProductDC() {
+	public Integer getProductDC() {
 		return productDC;
 	}
-	public void setProductDC(Double productDC) {
+	public void setProductDC(Integer productDC) {
 		this.productDC = productDC;
 	}
 	public Integer getProductStock() {
